requirements.txt
streamlit
akshare
talib-binary
pandas
numpy